requirements.txt
Flask==2.2.3
hikari==2.0.0.dev116
hikari-lightbulb==2.3.1
python-dotenv==0.21.1
requests==2.28.2
oauthlib==3.2.2
python-oauth2==1.1.1
requests-oauthlib==1.3.1
cryptography==39.0.1
supabase==1.0.0